Visual Studio 2017 fixes most of the warnings fixes

+++ src/argparse.c
@@ -387,11 +387,11 @@ argparse_usage (struct argparse *self)
         else if (options->type == ARGPARSE_OPT_STRING)
             pos += fprintf (stdout, "=<str>");
         if (pos <= usage_opts_width)
-            pad = usage_opts_width - pos;
+            pad = (int)(usage_opts_width - pos);
         else
         {
             fputc ('\n', stdout);
-            pad = usage_opts_width;
+            pad = (int)usage_opts_width;
         }
         fprintf (stdout, "%*s%s\n", pad + 2, "", options->help);
     }
